🚚 Rename UserList component files
